+++ backend/CulturalHeritage/src/main/java/tim2/CulturalHeritage/controller/RatingController.java
@@ -5,6 +5,9 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,10 +18,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import tim2.CulturalHeritage.dto.requestDTO.RatingRequestDTO;
-import tim2.CulturalHeritage.helper.RatingMappers.RatingRequestMapper;
-import tim2.CulturalHeritage.helper.RatingMappers.RatingResponseMapper;
+import tim2.CulturalHeritage.helper.ApiErrors;
+import tim2.CulturalHeritage.helper.RatingMapper;
+import tim2.CulturalHeritage.model.AuthenticatedUser;
+import tim2.CulturalHeritage.model.CulturalHeritage;
 import tim2.CulturalHeritage.model.Location;
 import tim2.CulturalHeritage.model.Rating;
+import tim2.CulturalHeritage.service.CulturalHeritageService;
 import tim2.CulturalHeritage.service.RatingService;
 
 import javax.validation.Valid;
@@ -30,9 +36,10 @@ public class RatingController {
     @Autowired
     private RatingService ratingService;
 
-    private RatingRequestMapper ratingRequestMapper = new RatingRequestMapper();
+    @Autowired
+    private CulturalHeritageService culturalHeritageService;
 
-    private RatingResponseMapper ratingResponseMapper = new RatingResponseMapper();
+    private RatingMapper ratingMapper = new RatingMapper();
 
     @GetMapping
     public ResponseEntity<List<Rating>> findAll() {
@@ -51,37 +58,57 @@ public class RatingController {
         }
     }
 
+    @PreAuthorize("hasRole('ROLE_USER')")
     @PostMapping
-    public ResponseEntity<?> add(@RequestBody RatingRequestDTO ratingRequest) {
+    public ResponseEntity<?> add(@Valid @RequestBody RatingRequestDTO ratingRequest, Errors errors) {
+        if (errors.hasErrors()) {
+            return new ResponseEntity(new ApiErrors(errors.getAllErrors()), HttpStatus.BAD_REQUEST);
+        }
+        AuthenticatedUser user = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
-        //ovde bi trebalo preuzeti id ulogovanog korisnika i setovati u novom rating objektu
         try {
-            Rating rating = ratingService.add(ratingRequestMapper.toEntity(ratingRequest));
+            Rating rating = ratingService.add(ratingMapper.toEntity(ratingRequest));
+            rating.setAuthenticatedUser(user);
+            CulturalHeritage ch = culturalHeritageService.findById(ratingRequest.getCulturalHeritageId());
+            rating.setCulturalHeritage(ch);
+            ratingService.add(rating);
 
-            return new ResponseEntity<>(ratingResponseMapper.toDto(rating), HttpStatus.CREATED);
+            return new ResponseEntity<>(ratingMapper.toDto(rating), HttpStatus.CREATED);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
+    @PreAuthorize("hasRole('ROLE_USER')")
     @PutMapping("/{id}")
-    public ResponseEntity<?> update(@PathVariable Long id, @RequestBody RatingRequestDTO ratingRequest) {
-
+    public ResponseEntity<?> update(@PathVariable Long id,@Valid @RequestBody RatingRequestDTO ratingRequest, Errors errors) {
+        if (errors.hasErrors()) {
+            return new ResponseEntity(new ApiErrors(errors.getAllErrors()), HttpStatus.BAD_REQUEST);
+        }
+        AuthenticatedUser user = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         try {
             Rating rating = ratingService.findById(id);
+            if (rating.getAuthenticatedUser().getId() != user.getId()) {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             rating.setGrade(ratingRequest.getGrade());
             ratingService.update(rating);
 
-            return new ResponseEntity<>(ratingResponseMapper.toDto(rating), HttpStatus.OK);
+            return new ResponseEntity<>(ratingMapper.toDto(rating), HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
+    @PreAuthorize("hasRole('ROLE_USER')")
     @DeleteMapping(path = "/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
-
+        AuthenticatedUser user = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         try {
+            Rating rating = ratingService.findById(id);
+            if (rating.getAuthenticatedUser().getId() != user.getId()) {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             ratingService.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
